refact: split API on slices
