+++ src/pages/SignUp/SignUp.jsx
@@ -4,7 +4,7 @@ import { EntryInput, Btn } from 'components';
 
 import { Link, useNavigate } from 'react-router-dom';
 import { useEffect, useState } from 'react';
-import { regNewUser, queryLogin, useSignupMutation, useLoginMutation } from 'services/API';
+import { useSignupMutation, useLoginMutation } from 'services/authAPISlice';
 import { useUser } from 'hooks';
 
 export const SignUp = () => {
@@ -16,6 +16,7 @@ export const SignUp = () => {
   const [password, setPassword] = useState('');
   const [passwordAgain, setPasswordAgain] = useState('');
   const [email, setEmail] = useState('');
+
   const [signup] = useSignupMutation();
   const [loginQuery] = useLoginMutation();
 
@@ -33,8 +34,10 @@ export const SignUp = () => {
       return;
     }
 
-    const { data: response, error: signupError } = await signup({ email, password });
-    console.log(response, 'Это ответ сервера на регистрацию');
+    const { data: response, error: signupError } = await signup({
+      email,
+      password,
+    });
 
     // Далее кривая обработка ошибок, ибо не умею
     if (signupError?.data) {
@@ -72,22 +75,6 @@ export const SignUp = () => {
     // Здесь убираю юзера в контекст
     login(userAndTokens[0]);
     navigate('/', { replace: true });
-
-    // regNewUser(email, password)
-    //   .then(() => queryLogin(email, password))
-    //   .then((response) => login(response))
-    //   .then(() => navigate('/'))
-    //   .catch((error) => {
-    //     console.warn(error);
-    //     if (error.response) {
-    //       console.warn(error.response.data);
-    //       const text = Object.values(error.response.data).join(' ');
-    //       setError(text);
-    //     } else {
-    //       console.log(error.request);
-    //       setError('Проблемы с сетью, проверьте подключение к сети интернет');
-    //     }
-    //   });
   };
 
   // Сбрасываем ошибку если пользователь меняет данные на форме
