+++ src/services/API.js
@@ -1,14 +1,9 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
-import axios from 'axios';
-import { setAccessToken, setRefreshToken } from 'store/UserSlice';
+import { setAccessToken } from 'store/UserSlice';
 
 const baseUrl = 'https://skypro-music-api.skyeng.tech/';
-const subURLs = {
-  all: 'catalog/track/all/',
-  allPlaylists: 'catalog/selection/',
-  favorites: 'catalog/track/favorite/all/',
-};
 
+// Базовый запрос, готовлю заголовки
 const baseQuery = fetchBaseQuery({
   baseUrl,
   prepareHeaders: (headers, { getState, endpoint }) => {
@@ -25,6 +20,7 @@ const baseQuery = fetchBaseQuery({
   },
 });
 
+// Обертка базового запроса: делает повторную авторизацию by JWT
 const baseQueryWithRefresh = async (args, api, extraOptions) => {
   let result = await baseQuery(args, api, extraOptions);
 
@@ -58,164 +54,10 @@ const baseQueryWithRefresh = async (args, api, extraOptions) => {
   console.log('Рефреш не произошел, необходимо перезагрузить страницу');
 };
 
+// Все эндпоинты лежат в своих слайсах
 export const musicServiceAPI = createApi({
   reducerPath: 'musicServiceAPI',
   baseQuery: baseQueryWithRefresh,
   tagTypes: ['Tracks'],
-
-  endpoints: (builder) => ({
-    // endpoint
-    refresh: builder.mutation({
-      query: () => ({
-        url: 'user/token/refresh/',
-        method: 'POST',
-        body: { refresh: localStorage.getItem('refreshToken') },
-      }),
-    }),
-
-    // endpoint В зависимости от переданного плейлиста загружает:
-    // все треки, конкретный плейлист или избранное
-    getTracks: builder.query({
-      query: (playlist = 'all') => {
-        let addURL = '';
-        // Если полученный аргумент можно привести к числу
-        if (!isNaN(Number(playlist))) addURL = `catalog/selection/${playlist}`;
-        else addURL = subURLs[playlist];
-        return addURL;
-      },
-
-      transformResponse: (data) => {
-        let tracks = data;
-        if (!Array.isArray(data)) tracks = data.items;
-        return tracks;
-      },
-
-      providesTags: (result) =>
-        // is result available?
-        result
-          ? // successful query
-            [
-              ...result.map(({ id }) => ({ type: 'Tracks', id })),
-              { type: 'Tracks', id: 'LIST' },
-            ]
-          : // an error occurred, but we still want to refetch this query when `{ type: 'Posts', id: 'LIST' }` is invalidated
-            [{ type: 'Tracks', id: 'LIST' }],
-    }),
-
-    // endpoint LIKE
-    likeTrack: builder.mutation({
-      query: (id) => ({
-        url: `catalog/track/${id}/favorite/`,
-        method: 'POST',
-      }),
-
-      invalidatesTags: (result, error, { id }) => [{ type: 'Tracks', id }],
-    }),
-
-    // endpoint DISLIKE
-    dislikeTrack: builder.mutation({
-      query: (id) => ({
-        url: `catalog/track/${id}/favorite/`,
-        method: 'DELETE',
-      }),
-
-      invalidatesTags: (result, error, { id }) => [{ type: 'Tracks', id }],
-    }),
-
-    // endpoint LOGIN
-    login: builder.mutation({
-      async queryFn({ email, password }, queryApi, _extraOptions, fetchWithBQ) {
-        const response = await Promise.all([
-          fetchWithBQ({
-            url: 'user/login/',
-            method: 'POST',
-            body: { email, password },
-          }),
-
-          fetchWithBQ({
-            url: 'user/token/',
-            method: 'POST',
-            body: { email, password },
-          }),
-        ]);
-
-        if (response[0].error) return response[0];
-
-        if (response[1].error) return response[1];
-
-        const userAndTokens = [response[0].data, response[1].data];
-        // Убираю токены в LS а также диспачу в store
-        localStorage.setItem('refreshToken', userAndTokens[1].refresh);
-        localStorage.setItem('accessToken', userAndTokens[1].access);
-        queryApi.dispatch(setRefreshToken(userAndTokens[1].refresh));
-        queryApi.dispatch(setAccessToken(userAndTokens[1].access));
-
-        return { data: userAndTokens };
-      },
-      extraOptions: { withoutRefresh: true },
-    }),
-    // endpoint Signup
-    signup: builder.mutation({
-      query: ({ email, password }) => ({
-        url: 'user/signup/',
-        method: 'POST',
-        body: { email, password, username: email },
-      }),
-      extraOptions: { withoutRefresh: true },
-    }),
-  }),
+  endpoints: (builder) => ({}),
 });
-
-export const {
-  useGetTracksQuery,
-  useLikeTrackMutation,
-  useDislikeTrackMutation,
-  useLoginMutation,
-  useSignupMutation,
-} = musicServiceAPI;
-
-export const regNewUser = (email, pass) => {
-  return axios
-    .post(baseUrl + 'user/signup/', {
-      email: email,
-      password: pass,
-      username: email,
-
-      headers: {
-        'content-type': 'application/json',
-      },
-    })
-    .then((response) => response.data);
-};
-
-// Логин и получение токенов одновременно, здесь же кладу refresh and
-// access Tokens в localStorage
-export const queryLogin = (email, password) => {
-  return Promise.all([
-    axios
-      .post(baseUrl + 'user/login/', {
-        email,
-        password,
-
-        headers: {
-          'content-type': 'application/json',
-        },
-      })
-      .then((response) => response.data),
-
-    axios
-      .post(baseUrl + 'user/token/', {
-        email: email,
-        password: password,
-
-        headers: {
-          'content-type': 'application/json',
-        },
-      })
-      .then(({ data }) => {
-        localStorage.setItem('refreshToken', data.refresh);
-        localStorage.setItem('accessToken', data.access);
-        return data;
-      }),
-  ]);
-};
